+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GST modes kwarg for optional model fitting
 - GST plotting for easier visualization of results
 - GST properties for easier access to results
+- GST for quantum instruments
 
 ### Changed